ima: Add appraisal support to IMA namespace
Enable each IMA namespace to appraise independently according to
their own appraisal policy.  Keys used to check the signature are
stored on a per-namespace IMA keyring.  Appraisal will only check
against keys on the namespace's keyring, but not recursively down
to the native.  Native is responsible for controlling what and when
keys can be loaded on a namespace's IMA keyring.

TODO:
  - IMA_NEW_FILE, IMA_DIGSIG, IMA_DIGSIG_REQUIRED should probably
    be added to the mask for ns_status->flags, or not? Possibly
    easrlier in the patch series already?

Signed-off-by: Yuqiong Sun <suny@us.ibm.com>
Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Signed-off-by: Mike Rapoport <rppt@linux.vnet.ibm.com>