ima: Enable per-namespace hash table
Provide a per-namespace hash table for storing measurement entries
that have been seen by the namespace.  This fixes the problem of the
current shared hash table, which will prevent the same measurements
from being stored on a namespace's measurement list even if it was
seen by native or other namespaces.

With the htable now in the ima_namespace structure we can now allow
namespace entries to be added to the htable.

Note: The #include linux/ima.h is needed on ppc64

Signed-off-by: Yuqiong Sun <suny@us.ibm.com>
Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>