ima: virtualize IMA's securityfs entries
Reorganize the securityfs entries of IMA so that the
securityfs entries of namespaces can be handled independently.

The directory '.ima' in the root dir contains the file entries for
init_ima_ns. It contains a namespaces directory under which each
namespace creates a directory with its own independent file entries
plus another namespaces directory that child containers would use
for putting their directories there.
The name of the created directory has the format ima:<inode num>.

The symlink 'ima' in the root dir serves to jump to the directory
of the current namespace.

In this patch we prevent updating of the policy from any other IMA
namespace than init_ima_ns.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>