fs: introduce SB_I_NEVER_TOO_REVEALING and use for sysfs
Sysfs cannot be mounted inside a container for as long as securityfs
is mounted to /sys/kernel/security due to it hiding files under the
security directory now.

Since we are cafeful enough when adding things to this directory we
should be fine to introduce a flag here that short-circuits the test
for whether a mount is too revealing. In case of sysfs, the mount would
never be too revealing.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>