ima: Get rid of IMA's dependency on securityfs for containers
Get rid of IMA's strict dependency on securityfs so it can be easier used
inside containers with a user namespace enabled. IMA can still be used
with securityfs but also only with sysfs. We do this by creating a symlink
in the security directory of sysfs that we can now use as a jumplink
similar to the one we still have in securityfs.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>