integrity: move measured_pcrs into struct ns_status
Move the measured_pcrs field from integrity_iint_cache into
ns_status struct so we can track the usage of PCRs on a per
namespace basis.

In some cases, when the measured_pcrs field is reset then it is
necessary to walk the list of ns_status connected to an iint and
reset it in all ns_status structures.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>