integrity: connect a list of ns_status to the integrity_iint_cache
In subsequent patches we will have to find all ns_status entries an iint
is being used in and reset flags there. To do this we now connect a list
of ns_status to the integrity_iint_cache and provide a reader-writer lock
in the integrity_iint_cache so we can lock access to it.

There are two cases for freeing:
- when the iint is freed: In this case we walk the list of ns_status
  entries and disconnect each ns_status on the list; we take the writer
  lock to protect access to the list
- when the ns_status is freed: In this case we only remove the ns_status
  from the list while also holding the iint's writer lock; to be able to
  grab the lock we have to have a pointer to the iint on the ns_status
  structure.

In the 1st case above an ns_status is not on the iint's list anymore and
therefore can be freed. The freeing is done lazily while walking the
rbtree with the rbtree's writer lock held.

We can run into the case that ns_status found in the RB tree has a
different iint associated with it for the same inode. In this case we
need to delete the ns_status structure and get a new one.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>