ima: prepare for current namespace != namespace where policy is in
Prepare process_measurement() and the functions it is calling to handle
the case where the IMA namespace that has the policy is different from
the one we are currently process measurements for. This allows us to
for example audit measurements due to the policy in the init_ima_ns
inside a child namespace.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>